Implement experimental UiPlugin interface

+++ src/painter.rs
@@ -160,10 +160,10 @@ impl Painter {
         let clip_min_y = pixels_per_point * clip_rect.min.y;
         let clip_max_x = pixels_per_point * clip_rect.max.x;
         let clip_max_y = pixels_per_point * clip_rect.max.y;
-        let clip_min_x = clamp(clip_min_x, 0.0..=width_pixels as f32);
-        let clip_min_y = clamp(clip_min_y, 0.0..=height_pixels as f32);
-        let clip_max_x = clamp(clip_max_x, clip_min_x..=width_pixels as f32);
-        let clip_max_y = clamp(clip_max_y, clip_min_y..=height_pixels as f32);
+        let clip_min_x = clamp(clip_min_x, 0.0..=width_pixels);
+        let clip_min_y = clamp(clip_min_y, 0.0..=height_pixels);
+        let clip_max_x = clamp(clip_max_x, clip_min_x..=width_pixels);
+        let clip_max_y = clamp(clip_max_y, clip_min_y..=height_pixels);
         let clip_min_x = clip_min_x.round() as u32;
         let clip_min_y = clip_min_y.round() as u32;
         let clip_max_x = clip_max_x.round() as u32;
